Move skin part into the skin namespace.
