+++ osu.Game.Rulesets.Karaoke.Tests/Screens/Skin/TestSceneConfigScreen.cs
@@ -4,7 +4,7 @@
 using osu.Game.Rulesets.Karaoke.Screens.Skin.Config;
 using osu.Game.Rulesets.Karaoke.Skinning;
 
-namespace osu.Game.Rulesets.Karaoke.Tests.Screens
+namespace osu.Game.Rulesets.Karaoke.Tests.Screens.Skin
 {
     public partial class TestSceneConfigScreen : KaraokeSkinEditorScreenTestScene<ConfigScreen>
     {
